+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- added new C API function `pineappl_grid_set_remapper`
+
 ### Changed
 
 - The command-line interface `pineappl` no longer prints both the differential